move filters to demodulator since that is where they are used
