+++ qcodes/instrument_drivers/AlazarTech/acq_controllers/demodulator.py
@@ -1,5 +1,48 @@
 import numpy as np
-import qcodes.instrument_drivers.AlazarTech.acq_helpers as helpers
+from scipy import signal
+
+def filter_win(rec, cutoff, sample_rate, numtaps, axis=-1):
+    """
+    low pass filter, returns filtered signal using FIR window
+    filter
+
+    Args:
+        rec: record to filter
+        cutoff: cutoff frequency
+        sample_rate: sampling rate
+        numtaps: number of frequency comppnents to use in the filer
+        axis: axis of record to apply filter along
+    """
+    nyq_rate = sample_rate / 2.
+    fir_coef = signal.firwin(numtaps, cutoff / nyq_rate)
+    filtered_rec = signal.lfilter(fir_coef, [1.0], rec, axis=axis)
+    return filtered_rec
+
+
+def filter_ls(rec, cutoff, sample_rate, numtaps, axis=-1):
+    """
+    low pass filter, returns filtered signal using FIR
+    least squared filter
+
+    Args:
+        rec: record to filter
+        cufoff: cutoff frequency
+        sample_rate: sampling rate
+        numtaps: number of frequency comppnents to use in the filer
+        axis: axis of record to apply filter along
+    """
+    raise NotImplementedError
+
+
+def filter_ham(rec, cutoff, sample_rate, numtaps):
+    raise NotImplementedError
+    # sample_rate = self.sample_rate
+    # nyq_rate = sample_rate / 2.
+    # fir_coef = signal.firwin(numtaps,
+    #                          cutoff / nyq_rate,
+    #                          window="hamming")
+    # filtered_rec = 2 * signal.lfilter(fir_coef, 1.0, rec)
+    # return filtered_rec
 
 class Demodulator:
 
@@ -66,23 +109,23 @@ class Demodulator:
         cutoff = self.active_channels[0]['demod_freq']/10
         # self.demod_freqs.get_max_demod_freq() / 10
         if self.filter_settings['filter'] == 0:
-            re_filtered = helpers.filter_win(re_mat, cutoff,
-                                             self.sample_rate,
-                                             self.filter_settings['numtaps'],
-                                             axis=-1)
-            im_filtered = helpers.filter_win(im_mat, cutoff,
-                                             self.sample_rate,
-                                             self.filter_settings['numtaps'],
-                                             axis=-1)
+            re_filtered = filter_win(re_mat, cutoff,
+                                     self.sample_rate,
+                                     self.filter_settings['numtaps'],
+                                     axis=-1)
+            im_filtered = filter_win(im_mat, cutoff,
+                                     self.sample_rate,
+                                     self.filter_settings['numtaps'],
+                                     axis=-1)
         elif self.filter_settings['filter'] == 1:
-            re_filtered = helpers.filter_ls(re_mat, cutoff,
-                                            self.sample_rate,
-                                            self.filter_settings['numtaps'],
-                                            axis=-1)
-            im_filtered = helpers.filter_ls(im_mat, cutoff,
-                                            self.sample_rate,
-                                            self.filter_settings['numtaps'],
-                                            axis=-1)
+            re_filtered = filter_ls(re_mat, cutoff,
+                                    self.sample_rate,
+                                    self.filter_settings['numtaps'],
+                                    axis=-1)
+            im_filtered = filter_ls(im_mat, cutoff,
+                                    self.sample_rate,
+                                    self.filter_settings['numtaps'],
+                                    axis=-1)
         elif self.filter_settings['filter'] == 2:
             re_filtered = re_mat
             im_filtered = im_mat
