+++ qcodes/instrument_drivers/AlazarTech/acq_helpers.py
@@ -1,49 +1,4 @@
 import numpy as np
-from scipy import signal
-
-
-def filter_win(rec, cutoff, sample_rate, numtaps, axis=-1):
-    """
-    low pass filter, returns filtered signal using FIR window
-    filter
-
-    Args:
-        rec: record to filter
-        cutoff: cutoff frequency
-        sample_rate: sampling rate
-        numtaps: number of frequency comppnents to use in the filer
-        axis: axis of record to apply filter along
-    """
-    nyq_rate = sample_rate / 2.
-    fir_coef = signal.firwin(numtaps, cutoff / nyq_rate)
-    filtered_rec = signal.lfilter(fir_coef, 1.0, rec, axis=axis)
-    return filtered_rec
-
-
-def filter_ls(rec, cutoff, sample_rate, numtaps, axis=-1):
-    """
-    low pass filter, returns filtered signal using FIR
-    least squared filter
-
-    Args:
-        rec: record to filter
-        cufoff: cutoff frequency
-        sample_rate: sampling rate
-        numtaps: number of frequency comppnents to use in the filer
-        axis: axis of record to apply filter along
-    """
-    raise NotImplementedError
-
-
-def filter_ham(rec, cutoff, sample_rate, numtaps):
-    raise NotImplementedError
-    # sample_rate = self.sample_rate
-    # nyq_rate = sample_rate / 2.
-    # fir_coef = signal.firwin(numtaps,
-    #                          cutoff / nyq_rate,
-    #                          window="hamming")
-    # filtered_rec = 2 * signal.lfilter(fir_coef, 1.0, rec)
-    # return filtered_rec
 
 
 def sample_to_volt_u12(raw_samples, bps, input_range_volts):
